LICENSE
Glozzom is licensed by Bipronath Saha. It's a template design and many have done this for their own work or something else. But this design which is deployed in Netlify is my own design that is completed by myself.  This is designed for working purposes not to use any commercial purpose. I, thereby, committed myself that I'll no longer use this template for any other projects.